#3563 VisitorInternal now also accumulates errors. Number conversion can now cause exception. Still prints warnings to stdout if conversion occurs.

+++ krLanguages/swiprolog/src/main/java/swiprolog/language/PrologExpression.java
@@ -161,8 +161,8 @@ public abstract class PrologExpression implements Expression {
 			return Long.toString(this.term.longValue());
 		}
 
-		if (this.term.isFloat()) {
-			return Float.toString(this.term.floatValue());
+		if (this.term.isFloat()) { // actually a double.
+			return Double.toString(this.term.doubleValue());
 		}
 
 		if (this.term.isCompound()) {
